Catch NoMethodErrors in FEDEX find_and_parse_rate method
As FEDEX is not always returning the service type for some rates
in the API currently used by ActiveShipping, some logic was
failing in our parsing code, namely trying to convert a missing
field into text would raise an error. This error would percolate
up to our main application and was unexpected by the exception
logic.

This fix catches the exceptions in ActiveShipping for all cases
in the find_and_parse method