#date_for should return a Date, not a DateTime
I believe this method is the source of a tricky bug in my application, and I suspect it is in others as well.

I discovered it while inspecting API responses from Stamps.com. As indicated by the method signature, API responses are expected to provide _calendar_ date information – any timezone information is discarded here, is correct. However, it returns an instance of DateTime, which includes time zone information. In this case that time zone is set to UTC (+0000) which is likely not correct.

If the resulting value is passed, for instance, to a Rails app that saves it in a timestamp field, it will be erroneously converted to a local time thus possibly changing the date. Changing this method to return a Date instance (as documented in the comment) seems more correct to me.